skb_tracking: move event to events mod
Also, move TrackingInfo event section.
