+++ src/module/skb_tracking/mod.rs
@@ -7,9 +7,6 @@
 pub(crate) mod skb_tracking;
 pub(crate) use skb_tracking::*;
 
-pub(crate) mod event;
-pub(crate) use event::*;
-
 mod tracking_hook {
     include!("bpf/.out/tracking_hook.rs");
 }
